project is done

+++ src/app/pages/user/user-dashboard/user-dashboard.component.ts
@@ -3,13 +3,13 @@ import { MatTableDataSource } from '@angular/material/table';
 import { Router } from '@angular/router';
 import { IConnectAdmin } from '../../../models/connect-admin';
 import { FontAwesomeService } from '../../../shared/icon';
-import { connectService } from '../../../services/connectService';
-import { historyService } from '../../../services/historyService';
+import { connectService } from '../../../services/pageService/connectService';
 import { authService } from '../../../services/authService';
 import { MatPaginator } from '@angular/material/paginator';
 import { MatDialog } from '@angular/material/dialog';
 import { ModalComponent } from './modal/modal.component';
 import { toastService } from '../../../shared/toast';
+import { historyService } from '../../../services/pageService/historyService';
 
 @Component({
   selector: 'app-user-dashboard',
@@ -22,8 +22,12 @@ export class UserDashboardComponent {
   dataSource = new MatTableDataSource<IConnectAdmin>([]);
   currentPage: number = 1;
   totalItems: number = 0;
-  userId: number = 0;
   previousPendingItems: { [id: number]: number } = {};
+
+  notificationsShown: Set<number> = new Set();
+  notifications: string[] = [];
+  hasNewNotification: boolean = false;
+  showNotifications: boolean = false;
 
   constructor(
     public faService: FontAwesomeService,
@@ -41,54 +45,77 @@ export class UserDashboardComponent {
     'portSource',
     'idLast',
     'portTo',
-    'pending',
+    'serviceType',
+    'connectType',
     'moderator',
+    'duration',
+    'status',
+    // 'timeExpired',
     'note',
     'updatedAt',
     'actions',
   ];
 
   ngOnInit() {
-    this.userId = this.authService.getUser().id;
     this.loadConnects(this.currentPage, 8);
     setInterval(() => {
       this.loadConnects(this.currentPage, 8);
-    }, 20000); // mini second
+    }, 20000);
   }
 
   ngAfterViewInit() {
     this.dataSource.paginator = this.paginator;
   }
 
-  async loadConnects(page: number, limit: number) {
-    const previousData = this.dataSource.data.slice();
-    const response = await this.connectService.getAllConnectByUserId(
-      page,
-      limit,
-      this.userId
-    );
-    this.dataSource.data = response.data.data;
-    this.currentPage = page;
-    this.totalItems = response.data.totalItems;
+  loadConnects(page: number, limit: number) {
+    this.connectService
+      .getAllConnectByUserId(page, limit, this.authService.currentUserValue.id)
+      .then((response: any) => {
+        const previousData = this.dataSource.data.slice();
 
-    this.dataSource.data.forEach((connect, index) => {
-      const previousConnect = previousData.find((c) => c.id === connect.id);
-      if (previousConnect) {
-        if (previousConnect.pending !== connect.pending) {
-          if (connect.pending === 0) {
-            this.toastr.success(
-              `Connection from ${connect.idFirst} to ${connect.idLast} has been opened.`,
-              'Nofication'
-            );
-          } else if (connect.pending === 1) {
-            this.toastr.warning(
-              `Connection from ${connect.idFirst} to ${connect.idLast} has been closed.`,
-              'Nofication'
-            );
+        this.currentPage = page;
+        this.dataSource.data = response.data;
+        this.totalItems = response.totalItems;
+
+        this.dataSource.data.forEach((connect: any) => {
+          const previousConnect = previousData.find((c) => c.id === connect.id);
+          if (previousConnect) {
+            if (previousConnect.status !== connect.status) {
+              if (connect.status === 0) {
+                const message = `Connection from ${connect.idFirst} to ${connect.idLast} has been opened.`;
+                this.notifications.push(message);
+                // this.toastr.success(message, 'Notification');
+                this.hasNewNotification = true;
+              } else if (connect.status === 1) {
+                const message = `Connection from ${connect.idFirst} to ${connect.idLast} has been closed.`;
+                this.notifications.push(message);
+                // this.toastr.warning(message, 'Notification');
+                this.hasNewNotification = true;
+              }
+            }
           }
-        }
-      }
-    });
+
+          // check time out
+          const timeExpired = new Date(connect.timeExpired);
+          const now = new Date();
+          const timeDiff = timeExpired.getTime() - now.getTime(); // time left (milisecond)
+          const warningTime = 1 * 60 * 1000; // milisecond
+
+          if (
+            timeDiff > 0 &&
+            timeDiff <= warningTime &&
+            !this.notificationsShown.has(connect.id)
+          ) {
+            const message = `Connection from ${connect.idFirst} to ${connect.idLast} is about to expire in 1 minute.`;
+            this.notifications.push(message);
+            this.hasNewNotification = true;
+            this.notificationsShown.add(connect.id);
+          }
+        });
+      })
+      .catch((error) => {
+        console.error('Error fetching connections:', error);
+      });
   }
 
   openDialog(element?: IConnectAdmin) {
@@ -111,9 +138,20 @@ export class UserDashboardComponent {
     this.openDialog(element);
   }
 
-  async handleSoftDeleteConnect(element: number) {
-    await this.connectService.deleteConnect(element);
-    this.toastr.success('Deleted successfully!', 'Notification');
+  async handleSendConnect(element: any) {
+    await this.connectService.sendConnect(element.id);
+    this.toastr.success('Connection was sent successfully!', 'Notification');
+    this.loadConnects(this.currentPage, 8);
+  }
+
+  async handleDeleteConnect(element: number) {
+    const confirmation = window.confirm(
+      'Are you sure you want to delete this connect?'
+    );
+    if (confirmation) {
+      await this.connectService.deleteConnect(element);
+      this.toastr.success('Deleted successfully!', 'Notification');
+    }
     this.loadConnects(this.currentPage, 8);
   }
 
@@ -125,5 +163,12 @@ export class UserDashboardComponent {
     this.currentPage = event.pageIndex + 1;
     const pageSize = event.pageSize;
     this.loadConnects(this.currentPage, pageSize);
+  }
+
+  toggleNofication() {
+    if (this.showNotifications) {
+      this.hasNewNotification = false;
+    }
+    this.showNotifications = !this.showNotifications;
   }
 }
